Liberally sprinkle #ifdef for protocol checking.
Use a lot of #ifdef to see what protocol versions are
available. Since it appears that the SSL_TXT_* macros
are consistent in their availability when there is
code to back up that protocol version, let's hang our
hat on that. I'm not sure if there is a better method
but there doesn't seem to be an easily documented way
to say if a particular protocol is supported by installed
library short of doing a horrible dlopen work. I'm not
paid enough to do that.