Optimize algorithm for server checking.
Instead of checking each and every cipher individually, connect
allowing all ciphers and then slowly prune the list of ciphers to
of the server preferred cipher. Once we fail to connect, we know
the server doesn't support any of the remaining ciphers. This
speeds up detection drastically as we only have check ciphers the
server actually supports. As a bonus, we also get the exact ordered
cipher list of the server.

No longer generate a STAILQ of all ciphers, we don't need it.
Manipulate the SSL_CTX instead of the SSL as it slightly changes
the behavior cipher list.